+++ src/Loaders/nfs4_loader.cpp
@@ -836,7 +836,7 @@ std::vector<TrackBlock> NFS4::ParseTRKModels(std::shared_ptr<TRACK> track) {
                 for (int k = 0; k < obj_polygon_block.nobj; k++) {
                     //TODO: Animated objects here, obj_polygon_block.types
                     // Keep track of unique textures in trackblock for later OpenGL bind
-                    std::set<short> minimal_texture_ids_set;
+                    std::set<unsigned int> minimal_texture_ids_set;
                     // Mesh Data
                     std::vector<unsigned int> vertex_indices;
                     std::vector<glm::vec2> uvs;
@@ -885,7 +885,7 @@ std::vector<TrackBlock> NFS4::ParseTRKModels(std::shared_ptr<TRACK> track) {
                         texture_indices.emplace_back(object_polys[p].texture);
                     }
                     // Get ordered list of unique texture id's present in block
-                    std::vector<short> texture_ids = TrackUtils::RemapTextureIDs(minimal_texture_ids_set, texture_indices);
+                    std::vector<unsigned int> texture_ids = TrackUtils::RemapTextureIDs(minimal_texture_ids_set, texture_indices);
                     current_track_block.objects.emplace_back(Entity(i,  (j + 1) * (k + 1), NFS_3, OBJ_POLY, Track(obj_verts, norms, uvs, texture_indices, vertex_indices, texture_ids, obj_shading_verts, trk_block_center)));
 
                 }
@@ -912,7 +912,7 @@ std::vector<TrackBlock> NFS4::ParseTRKModels(std::shared_ptr<TRACK> track) {
                             glm::vec4(((shading_data >> 16) & 0xFF) / 255.0f, ((shading_data >> 8) & 0xFF) / 255.0f,
                                       (shading_data & 0xFF) / 255.0f, ((shading_data >> 24) & 0xFF) / 255.0f));
                 }
-                std::set<short> minimal_texture_ids_set;
+                std::set<unsigned int> minimal_texture_ids_set;
                 std::vector<unsigned int> vertex_indices;
                 std::vector<glm::vec2> uvs;
                 std::vector<unsigned int> texture_indices;
@@ -958,13 +958,13 @@ std::vector<TrackBlock> NFS4::ParseTRKModels(std::shared_ptr<TRACK> track) {
                     texture_indices.emplace_back(poly.texture);
                 }
                 // Get ordered list of unique texture id's present in block
-                std::vector<short> texture_ids = TrackUtils::RemapTextureIDs(minimal_texture_ids_set, texture_indices);
+                std::vector<unsigned int> texture_ids = TrackUtils::RemapTextureIDs(minimal_texture_ids_set, texture_indices);
                 current_track_block.objects.emplace_back(Entity(i, l, NFS_3, XOBJ, Track(verts, norms, uvs, texture_indices, vertex_indices, texture_ids, xobj_shading_verts, trk_block_center)));
             }
         }
 
         // Keep track of unique textures in trackblock for later OpenGL bind
-        std::set<short> minimal_texture_ids_set;
+        std::set<unsigned int> minimal_texture_ids_set;
         // Mesh Data
         std::vector<unsigned int> vertex_indices;
         std::vector<glm::vec2> uvs;
@@ -1026,7 +1026,7 @@ std::vector<TrackBlock> NFS4::ParseTRKModels(std::shared_ptr<TRACK> track) {
                 texture_indices.emplace_back(poly_chunk[k].texture);
             }
             // Get ordered list of unique texture id's present in block
-            std::vector<short> texture_ids = TrackUtils::RemapTextureIDs(minimal_texture_ids_set, texture_indices);
+            std::vector<unsigned int> texture_ids = TrackUtils::RemapTextureIDs(minimal_texture_ids_set, texture_indices);
 
             if(chnk == 6){
                 current_track_block.lanes.emplace_back(Entity(i, -1, NFS_4, LANE, Track(verts, norms, uvs, texture_indices, vertex_indices, texture_ids, trk_block_shading_verts, trk_block_center)));
