+++ src/Scene/Track.cpp
@@ -7,7 +7,7 @@
 #include "Track.h"
 #include "../Util/Utils.h"
 
-Track::Track(std::vector<glm::vec3> verts, std::vector<glm::vec2> uvs, std::vector<unsigned int> texture_indices, std::vector<unsigned int> indices, std::vector<short> tex_ids,
+Track::Track(std::vector<glm::vec3> verts, std::vector<glm::vec2> uvs, std::vector<unsigned int> texture_indices, std::vector<unsigned int> indices, std::vector<unsigned int> tex_ids,
              std::vector<glm::vec4> shading_data, glm::vec3 center_position) : super("TrackMesh", verts, uvs, std::vector<glm::vec3>(), indices, true, center_position){
     m_texture_indices = texture_indices;
     shadingData = shading_data;
@@ -20,7 +20,7 @@ Track::Track(std::vector<glm::vec3> verts, std::vector<glm::vec2> uvs, std::vect
     ASSERT(genBuffers(), "Unable to generate GL Buffers for Track Model");
 }
 
-Track::Track(std::vector<glm::vec3> verts, std::vector<glm::vec3> norms, std::vector<glm::vec2> uvs, std::vector<unsigned int> texture_indices, std::vector<unsigned int> indices, std::vector<short> tex_ids,
+Track::Track(std::vector<glm::vec3> verts, std::vector<glm::vec3> norms, std::vector<glm::vec2> uvs, std::vector<unsigned int> texture_indices, std::vector<unsigned int> indices, std::vector<unsigned int> tex_ids,
              std::vector<glm::vec4> shading_data, glm::vec3 center_position) : super("TrackMesh", verts, uvs, norms, indices, true, center_position){
     m_texture_indices = texture_indices;
     shadingData = shading_data;
